+++ test/stream.js
@@ -38,3 +38,20 @@ metatests.test(
     }
   }
 );
+
+metatests.test('MemoryWritable handles custom sizeLimit', async test => {
+  for (const path of fixturesPaths) {
+    const originalReadable = fs.createReadStream(path);
+    const memoryStream = new common.MemoryWritable(40);
+    originalReadable.pipe(memoryStream);
+    const result = memoryStream.getData();
+    if (path.includes('empty')) {
+      await test.resolves(result, await readFile(path));
+    } else {
+      await test.rejects(
+        result,
+        new RangeError('size limit exceeded by 20 bytes')
+      );
+    }
+  }
+});